eg - added backend for account merge and beautified all pages
I created a new model named user which is just a placeholder for when we use login details to retrieve userdata. I modified the mentor select page to use the mentor model instead of the inline model I created, did the same for the mentorship page which now uses the "user" model instead of the inline model I had. lastly, I used visual studio document formatting to make the html structure easier to read in all of the pages.