Changes to DB schema, and user creation methods
A redundant identity field was removed from the user entity. The relationship between a user and address was changed to be one to one and have the user entity reference the address id via foreign key. This will help when creating a user and adding a respective address to the user.